Update`resourceVSphereDatacenterImport` to import folder
Updates `resourceVSphereDatacenterImport` to include the datacenter folder in  which the datacenter object may exist. If found in a datacenter folder, the leading "/" is removed and the inventory path is used for the `folder` attribute during the import. If the datacenter does not reside in a folder, the value is left empty during the import. #1374

Signed-off-by: Ryan Johnson <johnsonryan@vmware.com>