[vSAN] Support the configuration for vSAN performance service.
This change enables the configuration for vSAN performance service
via the .tf file. The three parameters supported are to enable/disable
vSAN performance service, the support for verbose mode and network
diagnostics mode. In this change, we will also support the
enablement for vSAN performance by default with vSAN performance
which aligns with the behavior for vSAN enablement UI.

The related documentation and API command line are also updated
in this changeset.