feat: add support for toggling vSAN unmap feature.
This change adds the feature to supporte the vSAN unmap feature. In
this code change, the vsan_unmap_enabled field is introduced, which
maps to the vSAN enable state. By enabling the feature, vSAN can
recognize the Trim/Unmap command from guestOS and relaim the storage
space that is mapped to a deleted vSAN object.