feat: add support for vSAN unmap (#1745)
This change adds the feature to support the vSAN unmap feature. In this code change, the `vsan_unmap_enabled` attribute is introduced, which maps to the vSAN enable state. By enabling the feature, vSAN can recognize the trim / unmap command from guestOS and reclaim the storage space that is mapped to a deleted vSAN object.