fix: race condition configuring vSAN and vSAN services (#1800)
When we are enabling vSAN and the vSAN related features, like
dedupe/compression or vSAN performance service, there could be a race condition
that these vSAN features cannot be properly enabled. On the vSphere task
dashboard, we will see the task failed with error message "The operation is not
supported on the object".

The root cause of the above issue is that we can see the enabling vSAN
operation and the enable vSAN feature operation are in two API calls to
backend, and they are not properly synchronized to avoid the race condition
that when the vSAN features are being enabled, vSAN is already in a stable
state of enabled.

This change fixes the above issue by setting up the proper API call sequence in
the vSAN and HA enable/disable workflow to ensure that the enable state can be
properly configured.