Feat: Add the vSAN DIT and HCI feature support
This change is to add the vSAN DIT (data-in-transit encryption)
and HCI mesh support. The configuration parameters include the
following parameters.

* DIT - enable flag and the rekey interval
* HCI mesh - the remote datastore IDs and the host IDs.

With the above two feature support, the users should be able
to configure the vSAN data-in-transit encryption and HCI
mesh feature at the same level of support of the vSphere UI.

Note: vSAN data-in-transit feature cannot be enabled with the HCI
mesh feature at the same time:
https://docs.vmware.com/en/VMware-vSphere/7.0/com.vmware.vsphere.vsan.doc/GUID-9113BBD6-5428-4287-9F61-C8C3EE51E07E.html.
This conflicts is called out in this code change as well.