Bugfix for issue 1939.
The code now correctly reads hardware version from the template and the terraform spec and passes along the higher of the two values to the environment browser.
This allows configuration of vm deployments from templates with OSes newer than what the default hardware compatibility version of the ESXi server allows.