Fixing cloning regression on DS cluster
Issue is that when vm is cloned on DS cluster the datastore_id field in  the
relocate spec holds invalid value which causes the vCenter to throw an
error. The fix was to restore previously implemented behavior to not
send relocate specs for the VM hard disks when it is cloned on DS
Cluster with exception when datastore_id is explicitly specified for the
Hard Disk.

Added TestAccResourceVSphereVirtualMachine_cloneOnDsCuster to cover the
case.

Fixes #2034

Signed-off-by: Vasil Atanasov <vasila@vmware.com>