[feat sr-iov] Adding memory_reservation_locked_to_max property to vsphere_virtual_macine schema
Fixes #2091
The  memory_reservation_locked_to_max property was missing from
vsphere_virtual_macine schema (both data and resource). The property of
the VM was automaticaly set to false if memory and memory_reservation
properties have different values. With this commit the property can be
set to true explicitly and it will be applied only if  memory and
memory_reservation properties have same values. If the
memory_reservation_locked_to_max is not explicitly set to true the
original behavior is preserved.

Signed-off-by: Vasil Atanasov <vasila@vmware.com>