[fix] SCSI IDs changing on machines built with 2.6
After the introduction of the SR-IOV feature network adapters due to
inconsistent check for changes in the `physical_function` attribute have
been recreated after clone since the check always returned `true`. The
result is that instead of updating existing NICs the relocate tack
started after clone was always deleting the existing NICs and creating
new ones. This was causing the new VM to be disconnected from network

Changed the check for changes in `physical_function` attribute to treat
nil and empty string equaly so missing `physical_function` attribute in
the device compared to empty string from the schema won't be cosnidered as changed.

Testing done: cloned VM from template with 2 nics and verified that
there is network connectivity. Also verified the output from lspci
command on the template VM and on the clone VM.

Fixes #2089

Signed-off-by: Vasil Atanasov <vasila@vmware.com>