fix: Differences between creation and import state
There are differences between the state in the r/vsphere_virtual_machine
when it is imported and when it is created/cloned. Fixed several of
them which maked sense for me.

- datastore_cluster_id - adding the property to the state when the VM is
  imported in order to keep the pervious behaviour.
- disk.label / disk.keep_on_remove - the logic was to assign a computed
  label in format "disk<number of the diks>" instead of the actual label
  added by the system and in such cases the disk was marked as
keep_on_remove. Changed this to set the actual label and not setting the
keep_on_remove property (defaulting to folse). If the label is not
present - keeping the old logic.

- Not sure how to procede with sata_controller_count - looks like this
  should be a computed proprty, but this is a breaking change which I
  would avoid introducung.

added sata_controller_count attribute  to ignore state attributes since
e2e tests are failing localy

Fixes #1706

Signed-off-by: Vasil Atanasov <vasila@vmware.com>