fix: `r/vsphere_virtual_machine` skip updating attributes if no diff
`GetDiffProcessorIfAttributesDefined` function ix missing a `HasChange`
check which is causing the provider to always update custom attributes
regardless if there is a diff or not.

This may cause problems in some cases, such as when you want to restrict
the permissions the terraform account has on vsphere. If the permissions
are locked down you'll end up getting a permission denied error.